refactor to use filter operator
