+++ servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/FilterPublisher.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018 Apple Inc. and the ServiceTalk project authors
+ * Copyright © 2021 Apple Inc. and the ServiceTalk project authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,30 +17,39 @@ package io.servicetalk.concurrent.api;
 
 import io.servicetalk.concurrent.internal.ConcurrentSubscription;
 
+import java.util.HashSet;
+import java.util.Set;
 import java.util.function.Predicate;
+import java.util.function.Supplier;
 import javax.annotation.Nullable;
 
 import static io.servicetalk.concurrent.internal.SubscriberUtils.checkDuplicateSubscription;
 import static java.util.Objects.requireNonNull;
 
-/**
- * As returned from {@link Publisher#filter(Predicate)}.
- *
- * @param <T> Type of items emitted by source {@link Publisher}
- */
 final class FilterPublisher<T> extends AbstractSynchronousPublisherOperator<T, T> {
-    private final Predicate<? super T> predicate;
+    private final Supplier<? extends Predicate<? super T>> filterSupplier;
 
-    FilterPublisher(Publisher<T> source, Predicate<? super T> predicate) {
+    FilterPublisher(final Publisher<T> source, final Supplier<? extends Predicate<? super T>> filterSupplier) {
         super(source);
-        this.predicate = requireNonNull(predicate);
+        this.filterSupplier = filterSupplier;
+    }
+
+    static <T> Supplier<? extends Predicate<? super T>> newDistinctSupplier() {
+        return () -> new Predicate<T>() {
+            private final Set<T> set = new HashSet<>();
+            @Override
+            public boolean test(final T t) {
+                return set.add(t);
+            }
+        };
     }
 
     @Override
-    public Subscriber<? super T> apply(Subscriber<? super T> subscriber) {
+    public Subscriber<? super T> apply(final Subscriber<? super T> subscriber) {
         return new Subscriber<T>() {
             @Nullable
             private Subscription subscription;
+            private final Predicate<? super T> predicate = requireNonNull(filterSupplier.get());
 
             @Override
             public void onSubscribe(Subscription s) {
@@ -52,12 +61,12 @@ final class FilterPublisher<T> extends AbstractSynchronousPublisherOperator<T, T
 
             @Override
             public void onNext(T t) {
-                // If Predicate.test(...) throws we just propagate it to the caller which is responsible to terminate
+                // If predicate or keySelector throws we propagate to the caller which is responsible to terminate
                 // its subscriber and cancel the subscription.
                 if (predicate.test(t)) {
                     subscriber.onNext(t);
                 } else {
-                    assert subscription != null : "Subscription can not be null in onNext.";
+                    assert subscription != null;
                     subscription.request(1); // Since we filtered one item.
                 }
             }
